fix(cli): rename dev ui constants to drop the duplicate `constants` basename
CI's `tsc` build deterministically failed to resolve `#utils/constants.js`
from one importer (TS2307) once the dev-views import graph was reshuffled — a
TypeScript module-resolution quirk triggered by two `constants.ts` files both
reachable via `#` aliases (`#utils/constants.js` and
`#views/dev/utils/constants.js`). It reproduced only on CI's filesystem
ordering, never locally; tsconfig `paths` / package `imports` tweaks had no
effect (tsc resolves these via `paths`, keyed by full specifier).

Rename the dev module to `ui_constants.ts` so no two `#`-reachable modules
share the `constants` basename. Reverts the earlier paths/imports experiments.
