+++ sdk/cli/tsconfig.json
@@ -17,11 +17,6 @@
     "resolveJsonModule": true,
     "baseUrl": ".",
     "paths": {
-      "#api/*": ["src/api/*"],
-      "#services/*": ["src/services/*"],
-      "#types/*": ["src/types/*"],
-      "#utils/*": ["src/utils/*"],
-      "#views/*": ["src/views/*"],
       "#*": ["src/*"]
     }
   },
